move wasm client tests
